refactor: MyAccountBox - MyRefrigeratorBox 리팩토링
